Fix : exponenter method Java
Bug : Initialized 'mul' variable as int and assigning original 'num' to 'mul' which is assigning double variable to an int variable causing data loss and syntax error.

Fix : Initialized 'mul' variable with double datatype.

+++ postfix_solver.java
@@ -284,7 +284,7 @@ public class postfix_solver
 
     public static double exponenter (double num, int exponent)
     {
-        int mul = num;
+        double mul = num;
         if (exponent == 0)
         {
             return 1;
